HuyDatPhong

+++ WebApplication1/Controllers/UserController.cs
@@ -517,53 +517,57 @@ namespace WebApplication1.Controllers
             {
                 return Redirect("DangNhap");
             }
+            DateTime dateHomNay = DateTime.Now;
             var lst = db.DATPHONGs.Where(dp => dp.MaKH == user.MaKH)
-                .Join(db.PHONGs, dp => dp.MaPH, p => p.MaPH, (dp, p) => new { dp, p })
-                .Join(db.LOAIPHONGs, dp_p => dp_p.p.MaLoai, lp => lp.MaLoai, (dp_p, lp) => new
+                .Join(db.PHONGs, dp => dp.MaPH, p => p.MaPH, (dp, p) => new
                 {
-                    MaDatPhong = dp_p.dp.MaDP,
-                    TenPhong = dp_p.p.SoPH,
-                    LoaiPhong = lp.TenLoai,
-                    NgayDat = dp_p.dp.NgayDat,
-                    NgayNhan = dp_p.dp.NgayNhan,
-                    NgayTra = dp_p.dp.NgayTra,
-                    DichVu = dp_p.dp.MaDV,
-                    ThanhTien = dp_p.dp.DonGia
+                    MaDatPhong = dp.MaDP,
+                    TenPhong = p.SoPH,
+                    NgayDat = dp.NgayDat,
+                    NgayNhan = dp.NgayNhan,
+                    NgayTra = dp.NgayTra,
+                    DichVu = dp.MaDV,
+                    ThanhTien = dp.DonGia
                 }).AsEnumerable().Select(m => 
                     new LichSuView()
                     {
                         MaDatPhong = m.MaDatPhong,
                         TenPhong = m.TenPhong,
-                        LoaiPhong = m.LoaiPhong,
                         NgayDat = m.NgayDat.Value.ToString("dd/MM/yyyy"),
                         NgayNhan = m.NgayNhan.Value.ToString("dd/MM/yyyy"),
                         NgayTra = m.NgayTra.Value.ToString("dd/MM/yyyy"),
                         DichVu = m.DichVu,
                         ThanhTien = m.ThanhTien,
-                        CoTheHuy = m.NgayNhan > DateTime.Now ? true : false
+                        CoTheHuy = m.NgayNhan > dateHomNay ? true : false
                     }
                 ).ToList();
             return View(lst);
         }
 
-        [HttpPost]
+        [HttpGet]
         public ActionResult HuyDatPhong(string maDP)
         {
+            if (string.IsNullOrEmpty(maDP))
+            {
+                return HttpNotFound();
+            }
+
             var room = db.DATPHONGs.FirstOrDefault(s => s.MaDP == maDP);
             if (room == null)
             {
-                return HttpNotFound(); // hoặc trả về thông báo lỗi tùy ý
+                return HttpNotFound();
             }
 
-            else
-            {
-                db.DATPHONGs.Remove(room);
-                db.SaveChanges();
-            }
-            return View();
+            // Xóa bản ghi đặt phòng
+            db.DATPHONGs.Remove(room);
+            db.SaveChanges();
+
+            // Điều hướng lại trang lịch sử
+            return RedirectToAction("LichSuDatPhong");
         }
 
-        
+
+
     }
 
 }
